+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `plot_data_at_multipactor_thresholds`
   - `data_for_susceptibility`
   - `plot_instruments_y_vs_instrument_x`
+- `Instrument` methods:
+  - `values_at_barriers`
+  - `values_at_barriers_fully_conditioned`
+  - `plot_vs_time`
 
 
 ## [1.5.0] 2024-03-02